Eric's Big Adventure
This commit addresses the greater part of Eric's v1 feature-parity work,
formerly the `eac-dev` branch.

This commit has not yet been DM reviewed, but does pass linters, is mypy
clean, and has a functional test suite -- in light of this, the team has
chosen to conduct further review prep work incrementally on `main`.